Struct fetcher::clients::json_client::JsonClient

pub struct JsonClient<'a> {
    // some fields omitted
}

A Client that is backed by data from JSON files.

note: The files are read every time a Client function is called

Methods

impl<'a> JsonClient<'a>

fn new(series_path: &'a str, providers_path: &'a str, base_providers_path: &'a str) -> JsonClient<'a>

Create JsonClient with paths pointing directly to each file used for backing data.